+++ tests/consumers/browser.mjs
@@ -24,5 +24,6 @@ void client.statements;
 void client.webhooks;
 void acquiringClient.invoices;
 void acquiringClient.merchant;
+void acquiringClient.statements;
 void acquiringClient.webhooks;
 void verifyAcquiringWebhookSignature;